// frontend/src/app/core/models/Tramite/VoucherRequestModel.ts
export class VoucherRequestModel {
    tramiteId: string;
    tramiteReqId: string;
    tramiteReqRefId: string;
    tipoDocumento: string;
    numeroDocumento: string;
    codigoTributo: string;
    numeroSecuencia: string;
    fechaMovimiento: string;
    codigoOficina: string;
    total: number;
    codigoUsuario: string;

    // Datos Generales
    codigoProc: string;
    descProc: string;
}